refactor error styling
